Test onInputChange events on a per-character basis.
In @testing-library/user-event >= ^11, the way type() works has changed.
As opposed to firing a single event for the full supplied string, it
types in one character at a time, very much like how an actual user does
it. Because we don't supply the current state as props, the target
component really only knows about the current character being typed.

To solve this, we just use Jest's mocks to track each character typed.