Remove warning from Gallery component.
Decision was made to remove the check on children since there isn’t
really a strong reason. It’s the responsibility of the caller to make
sure they pass the right attributes. Also remove the automatic
generation of the alt attribute since it doesn’t make sense in most
contexts.

https://glints.atlassian.net/browse/SS-777